Correect a known IMRPhenomD bug in the way t=0 is set at an internally defined peak frequency.
There was a missing term in the FD phase derivative.
The bug is quite mild since the mapping to TD and the definition of
fpeak are approximate anyway. The correction does reduce a bit the
dispersion of peak times when computing the IFFT.
Note: this bug is known but nor corrected in LAL yet.